Fix, for android

+++ storm/Thread.cpp
@@ -14,6 +14,6 @@ uintptr_t SGetCurrentThreadId() {
 #endif
 
 #if defined(WHOA_SYSTEM_MAC) || defined(WHOA_SYSTEM_LINUX)
-    return reinterpret_cast<long unsigned int>(pthread_self());
+    return reinterpret_cast</*long unsigned int*/ intptr_t>(pthread_self());
 #endif
 }
